pyproject.toml
[project]
name = "x-searcher"
version = "1.0.0"
description = ""
authors = [
    {name = "tkc310",email = "takashisato.develop@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "twikit (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
